requirements.txt
av==15.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
coloredlogs==15.0.1
ctranslate2==4.6.0
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
openai-whisper==20250625
packaging==25.0
protobuf==6.32.1
pydub==0.25.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
six==1.17.0
sympy==1.14.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Flask==3.0.0
Flask-CORS==4.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
requests==2.31.0
python-multipart==0.0.6
Werkzeug==3.0.1
gunicorn==21.2.0